Fix bug where trying to cast from int to long

+++ src/presentation/dispatcher/ViewDeckDispatcher.java
@@ -28,7 +28,7 @@ public class ViewDeckDispatcher extends AbstractDispatcher {
 			if(checkIfLoggedIn(myRequest)) {
 				
 				String[] requestPath = myRequest.getServletPath().trim().split("/");
-				int deckID = Integer.parseInt(requestPath[requestPath.length-1]);
+				Long deckID = Long.parseLong(requestPath[requestPath.length-1]);
 				myHelper.setRequestAttribute("deckID", deckID);
 				
 				ViewDeckCommand viewDeck = new ViewDeckCommand(myHelper);
